Reorder change callback, fix mixin names
Boo validateMixinNames doesn't let me do the inner class thing
