+++ fabric-game-rule-api-v1/src/main/resources/fabric-game-rule-api-v1.mixins.json
@@ -6,8 +6,7 @@
     "GameRuleCommandAccessor",
     "GameRuleCommandVisitorMixin",
     "GameRuleMixin",
-    "GameRuleRpcDispatcherMixin",
-    "GameRuleRpcDispatcherMixin$class_12254Mixin",
+    "GameRuleRpcDispatcherclass_12254Mixin",
     "MinecraftServerMixin"
   ],
   "injectors": {
